Work towards lattice files for iDMRG, the basic structure is now well defined: the UnitCell defines a (usually small) list of LatticeSite's.  Local operators are indexed as Op[i], for zero-based indices i.  The UnitCell also allows the definition of global operators that have support over the entire UnitCell.  All of these operators also define a Jordan-Wigner string that can be used to lift the operator to a complete lattice.  As a special case, for UnitCell's of a single site, the operator parser will allow the [0] to be neglected.  It is therefore not valid to add a UnitCell operator of the same name as a local operator, if the unit cell is 1 site.  Operators defined in the UnitCell are allowed to have support on some multiple of unit cells.
An InfiniteLattice is defined over a single repeated UnitCell.  Unit cell operators are 
then referenced as Op(i), so eg a local operator could be referred to as Op(i)[j], for 
the j'th LatticeSite within the i'th unit cell.  For unit cell operators that have 
support over more than one unit cell, the (i) index refers to the left-most unit cell 
that has non-trivial support.

The InfiniteLattice also allows the definition of operators with infinite support.  
Currently InfiniteMPO is a typedef for TriangularMPO, but it is intended that we will be 
able to add infinite product operators as well (as type ProductMPO?).  Again, the unit 
cell of an InfiniteMPO could be a multiple of the size of the basic UnitCell. The parser 
will refer to these operators as Op with no brackets.  So it would be possible, for 
example, to define a non-trival unit cell with local operators S[0], S[1], ....  The 
UnitCell can then define an operator S = S[0]+S[1]+...  The InfiniteLattice can then 
define a TriangularMPO that is the sum of all of the unit cell S operators.  For the 
infinite wavefunction, these operators would be referred to as "S[i](j)" (S operator at 
the i'th LatticeSite of the j'th UnitCell), "S(j)" (total S operator of the j'th unit 
cell), and "S" (total S on the entire lattice).

It would be useful to define something like a UnitCellOperator, that acts similarly to 
the old OperatorAtSite. This is not yet added.