version number support for the FileSystem (for objects that don't have their own version number), and lanczos stopping criteria
Modified the lanczos stopping criteria to use ||A*x - theta*x|| / d, where d is the spectral diameter (estimated by
the difference between the highest and lowest Ritz eigenvalues).  This criteria is both scale- and shift-invariant.

Also added a bit of a hack, to get access to the metadata version number of a pheap FileSystem object.  This is done via the new
version() member of an ipheapstream.  So to use it,

   PHeapFileSystem::ipheapstream* S = dynamic_cast<PHeapFileSystem::ipheapstream*>(&in);
   if (S && S->version() == 1)
   {
      // old version
   }
   else
   {
      // new version
   }

Simultaneously, the FileSystem metadata version number has been increased from 1 to 2 (without changing the on-disk format).