beginning of C++-11 simplifications, strict pheap version checking
added ExpectedVersion to pheap, so that by default an attempt to open an old format file will fail.
Programs that are designed to read old files can call SetExpectedVersion(n) to allow reading
a specific version.

LinearAlgebra fix to make operator* for DiagonalMatrix work.

Started on some C++-11 simplifications for LinearAlgebra.  The extent of the current work is
a proof-of-principle, use a template type alias instead of the macro concept declarations.
This appears to work quite well, eg, instead of DIAGONAL_MATRIX(V, T), use DiagonalMatrixI<V,T>

Rearranged the Matrix concept heirachy: eliminated Rectangular, and made SparseMatrix an injective matrix.
In fact, injective isn't actually a distinct concept, we can probably remove it completely.

Likewise, the Vector concepts could have the LocalVector as a generic (non-injective), and the
Compressed can be an injective type.  On the other hand, there is nothing wrong with the vector
concepts as they are now.

Modified the Structure format for IrredTensor, to be a traits-like class that contains
a templated alias.  Two storage traits are currently defined, the DefaultStorage uses a
SparseMatrix, and DiagonalStorage uses DiagonalMatrix.  The RealDiagonalOperator now uses
the DiagonalStorage.  This modifies the on-disk format for wavefunctions, so the version
of CanonicalWavefunctionBase has been bumped to 4.

Adjusted some DEBUG_ checks, since it seems we had implicitly assumed that we could stream
a std::vector to an ostream?!  but we can't in GCC C++-11 mode.

It is unclear whether rvalue references are able to simplify the proxy logic.

Removed some obsolete files (README and TODO)

Changed DiagonalMatrix GetElement for the case where the element_type doesn't have a zero value,
previous policy was to return a value_with_zero<element_type>, but this is rather cumbersome
since we can't use it as an l-value, and it makes a copy of the element.  Now we simply
assert that the element is on the diagonal and return it directly.