major progress on numerical stability of idmrg wavefunctions.
still lots of debug stuff in there.  The big mystery now is the numerical accuracy
of the initial residual.  If we have a well-converged wavefunction, even if it has
a small number of states (so that the overall variance might be big), we ought to
have a very good eigenstate of the local Hamiltonian.  That is, find the left and
right Hamiltonian block operators, and the initial wavefunction should already be a good
eigenstate.  But it isn't - for practical examples the initial residual norm (which is
the square root of the variance) is of order 1E-5 - 1E-6.  This is crazy, because
eg for ITF or spin-1 chain, one can easily get an absolute variance ~ 1E-12, so
we have the contradiction that the variance in the restricted Hilbert space is larger
than the exact variance.

Other than that, the orthogonalization procedure is very stable.  Issue is when left
eigenvector has small components - which only happens in pathological cases - but we can either keep the states
(and have some small elements in the density matrix - has no effect on physical observables but should
be possible to handle) or remove them from the basis.  Both cases should be implemented.

Also fixed bug in svd for the 'full' case.  Generally well tested now with linearalgebra/test/testsvd.cpp