Adds --expression function to mp-aux-algebra
This is somewhat incomplete in the sense that it adds a parser for MatrixOperator, so only works
for operators that are 1-dimensional at the unit cell boundary.  This could be extended
to a general StateComponent expression parser, but that is a lot more work.

Adds exp() function for MatrixOperator

Add RyOld operator to spin-tri-yc model