Added conditional operator support for OperatorDescriptions, and pass UnitCell by pointer to InfiniteLattice
This is an optional 3rd or 4th parameter to the add_operators()(Name, Description, [Conditional], [Function]) function
The third parameter is a string, if this is not empty then the operator is treated as a conditional operator
and the condition is listed in the help string for the model.  Additionally, a 4th parameter is a function,
anything that can be represented in a std::function<bool()>, which is invoked to test whether the function should be
defined and print a warning if not (and vice versa).  This 4th parameter is optional, if it is absent then no warning
will be emitted if the operator is not defined.

The 4th parameter, if it is specified, will typically be a lambda function that takes some parameter by reference (since
we need the value at the point where the lambda is invoked, not when it is constructed), and return true if the operator
should be defined.  For example,

      OpDescriptions.add_operators()
	 ("StagS"      , "staggered magnetization", "width even", [&w]()->bool {return w%2 == 0;})
         ;

where w is a parameter to the model that is set on the command line.

The UnitCell can now optionally be passed to the InfiniteLattice by pointer.  If this is done, then it is assumed
that the UnitCell has a longer lifetime than the InfiniteLattice.  This is generally true if they are both constructed
in the same scope.  This now allows modifications to the unit cell after the Lattice has been constructed.
This requires updating all models, since there is no longer a conversion ctor that takes
a non-rvalue UnitCell.  The options are:

1. pass the unit cell by pointer.

      LatticeSite Site = SpinSU2(Spin);
      UnitCell Cell = repeat(Site, w);
      InfiniteLattice Lattice(&Cell);

2. pass the unit cell from a temporary

      LatticeSite Site = SpinSU2(Spin);
      InfiniteLattice Lattice(repeat(Site, w));

3. 'move' the unit cell (and destroy the old object in the process)

      LatticeSite Site = SpinSU2(Spin);
      UnitCell Cell = repeat(Site, w);
      InfiniteLattice Lattice(std::move(Cell));

This 3rd option has no advantages over option 2, since now 'Cell' is an invalid object.  In most cases, option 1 is
what we want.