Added an amplitude term to InfiniteWavefunctionLeft.
The idea is to allow an amplitude per unit cell for infinite wavefunctions.  The main use case is imaginary time evolution, where the amplitude of the wavefunction changes as exp[-*Beta*E], and we want to track this amplitude.

Support for the amplitude is not yet complete.  For most tools, a decision needs to be made as to how to support the amplitude.

mp-ioverlap incorporates the amplitude into the result.

mp-itebd has options --normalize and --nonormalize.  The default is that real time evolution is normalized, and if there is any imaginary component to the timestep, it is not normalized (i.e. the amplitude tracks the changing amplitude of the wavefunction).  The default can be overridden by the --normalize or --nonormalize options.

mp-info displays the normalization.

mp-expectation doesn't use the normalization.  This is the correct choice - the only meaningful interpretation of an expectation value is <psi|Op|psi> / <psi|psi>, so the amplitude cancels out.

mp-imoments - similar situation to mp-expectation.

mp-idmrg doesn't use the normalization.  This means it will default to 1.0 when a wavefunction is created, and running mp-idmrg on an existing wavefunction doesn't change the normalization.

Some tools could be extended to use the amplitude:

* mp-irepeat needs checking to see if it does the right thing.

* mp-idivide will definitely handle the amplitude incorrectly.

* mp-norm, mp-normalize, mp-scale could be extended to manipulate the amplitude, in the obvious ways.

* mp-iapply could be generalised to apply any kind of periodic operator, eg a triangular MPO or a ProductMPO that doesn't result in a normalized wavefunction.