Various fixes and improvements
Add constructors for InfiniteWavefunctionLeft/Right which return the unitary
transforming the old edge basis to the new one.
mp-ibc-create: Ensure the window joins properly to the right boundary.
mp-ibc-join: Ensure the normalization of the output state.
mp-ibc-apply: Add the ability to apply a string operator to the left
semi-infinite boundary.