filegrid function works.
filegrid(f,x,y,z) reads the file named "f", and returns the element at array index (x,y,z).
The format of the file is the same as gnuplot 3D format.  The X direction is the rows.  The Y direction is the columns.  The Z direction is blocks separated by an empty line.

Example:

$ cat testfile
0 2
1 3

10 30
20 40
$ mp-expectation psi lat:"filegrid(testfile,0.5,0.5,0.5) *I(0)"
13.25

Currently the parsing of the filename is a bit primitive, it reads everything between the opening bracket and the comma, so extra spaces etc cause it to fail.

The file can contain numbers or an arbitrary expression, including an MPO.  Currently function arguments don't work properly if filegrid is used in a function defintion and it refers to arguments passed into the function.

The elements of the file are separated by whitespace.  To use expressions that contain spaces, you can use quotes, or backslash escapes.  For example:

$ cat testfile2
"2 * I(0)"
"4 * I(0)
$ ./mp-expectation psi lat:"filegrid(testfile2,0.5)"
3