EA: transform PsiRight to right canonical form initially
This change is necessary for us to save the wavefunctions properly. But it also
requires completely refactoring the intializers for HEff. I had initially two
initializers, one where PsiLeft and PsiRight were the same wavefunction and one
where they were strictly different. In this refactored version, there is only
one initializer, and it explicitly checks the overlap of PsiLeft/Right to see
if they are orthogonal. This should be okay, but it doesn't work at the moment
for states like the AKLT state, where the GMRES solver gets stuck when trying
to calculate the right Hamiltonian environment for PsiRight to find the bond
energy, but this bug probably needs to be fixed elsewhere anyway.