Orthogonalization seems to be working well now.
We now retry the eigensolver if the left or right transfer matrix eigenvectors are not sufficiently Hermitian.  We give up after 10 attempts (which never happend in testing) and continue anyway.

In tests of the Ising model through an imaginary time transition, the orthogonalization remains stable until way after roundoff errors have broken the symmetry, which appears as
a non-injective state, but seems to be two copies of the same state?  Perhaps that is from systematic error in the Trotter decomposition?  The density matrix is 2-fold degenerate.

This dramatically reduces the number of warnings that appear when orthogonalizing states.  The warnings are still in the code, but they aren't triggered very often.

We need some criteria to know when the wavefunction is numerically non-injective.  I think having a next-leading eigenvalue smaller than 1E-8 is about right, certainly it seems
the correct scale is sqrt(eps).