Initial implementation of mp-dmrg-ee (for finite size DMRG).
Issues:
Default value of --env-states.  Should this be multiplicative?
To get good convergence starting from a product state, it is required to force keeping zero weight states in the density matrix, which means setting
the -r parameter to something smaller than 0.  Eg

mp-dmrg-ee -H lat:H_J1 -w psi -m 10..100x10 -r -1

A way around this problem would be to use the 3S algorithm to add states to the basis, if we run out of states to keep that have weight > 0.
Eg, use 3S projected into the null space, and use that to choose the additional states to keep.  But simply keeping states of zero weight seems to work OK,
certainly when the number of additional environment states is not too big (eg if the environment states is exactly equal to the number of states we want
to keep, then we are not doing a truncation at all, and the additional states are already selected using 3S).  Or instead of doing full 3S, keep the
weight matrix from the environment expansion, and when we do the truncation project the weight matrix into the null space and use that to select the
additional states.  This has the disadvantage that the states are selected based on old information, rather than the recently obtained wavefunction.

But since we don't add environment states that have zero weight (except the per-quantum-sector minimum), perhaps this really isn't a problem.  Simply
we should set the default value of -r to be less than zero.

Even with --env-states 3, this converges faster than 3S (although strangely is slightly slower)