Changing local Hilbert spaces (#56)
* Some work on finite DMRG, and add the mp-change-lattice tool.

Operators that change the local Hilbert space need some additional thought and design work.
Operators, including MPO's, are allowed to change the local Hibert space.  The intention behind this is to
allow operators to map between different Hilbert spaces, eg to change the boson number cutoff, also
Gutzwiller projectors between fermions and spins, and other examples such as swap and translation operators
on non-homogeneous lattices.  This needs a bit more thought however, because currently when we create an MPO from
a local operator, it gets identity operators on all other sites except the active site, which means it isn't possible
to construct operators such as projection operators that act on more than 1 site (or in some cases we can, but we need to
be strict about the order of operations).

A lot of this will be simplified once we can eliminate JW strings, because that allows the possibility to have MPO's that
have support over a sparse (and disjoint) set of sites.  These MPO's could be multiplied together, and the Hilbert spaces only
need to match on the sites that have non-trivial support.  But for the time being it isn't really possible to make this kind of
operator.  As a work-around, which will probably still be useful once we can do more general projections, there is a tool
mp-change-lattice, which maps the local Hilbert space of one lattice to another.  The main use for this is to change the
boson number cutoff for bosonic models, but it has other potential uses too.

There are some miscellaneous changes that facilitate MPO's with Basis1() != Basis2(), which are harmess but not actually used
by the mp-change-lattice tool.

This commit also adds the --no-keep-list option to mp-dmrg.  This is needed to fix the number of states at exactly m, if that is desired.