EE with random environments works shockingly well!
In this experiment, we force the environment states to be selected from the Haar measure (subject to available quantum number sectors) by
forcing the mixing matrix to be N x 0, prior to the SVD.  The SVD properly handles this case and returns Haar-randomized left singular vectors.

Next step is to allow this properly in the code, and do some more testing.