mp-ibc-tdvp: Add ‘evolution windows’
Only a subset of the full IBC window (the evolution window) is evolved: this
allows us to expand the at the edges of the evolution window.

This contains the ‘comoving window’ scheme as well, where only a fixed width
window is evolved, which moves along with the wavefront.

This also fixes the fidelity loss window expansion criterion to use the actual
fidelity, and not the Frobenius norm, and removes the LambdaDiff criterion,
which is no longer needed.

This removes the UCExpand option to expand the window by unit cells rather than
by single sites: this was the original window expansion method before
single-site expansion. If unit-cell expansion is needed again in the future, it
can probably be done in terms of the current single-site expansion code.