Fix write-after-free bug in pheap.
In the pheap object, when they are deallocated there is a mutex that protects the reference count.  When the reference count hits zero, the pheap object is self-deleted (delete this).
But the reference count is protected by a mutex, so we need to ensure that we drop the mutex *before* we delete this.  Probably there isn't any need to release the mutex since
it is about to get deallocated, but re-orderering the operations properly is just as easy.