Stability improvements: ensure that the expansion vectors are orthogonal to the kept states when using range finding / rsvd by removing the projection a second time, after the QR.
Rewrote the svd pre-expansion to use the same approach as other pre-expansion, i.e. do the SVD in the null space and add the expansion vectors to the kept states.  The original svd algorithm just did an SVD of H|psi>.  The svd now agrees quite closely (but not exactly) with rsvd with maximal oversampling.