Add ExtraPerSector to OversamplingInfo
When oversampling, we can add extra states per sector even if the ExtraStatesPerSector is zero for the final RSVD.
The oversampling states will simply be in the mix to be kept based on their singular values.  Without this step, the RSVD
has very poor performance with ExtraStatesPerSector=0, but setting ExtraStatesPerSector=1 can blow out the number of
used quantum number sectors.  Adding extra states per sector just in the oversampling phase gets the best of both worlds.

With this change, ExtraStatesPerSector is only relevant for 'range' and 'random' algorithms.