Reorganization of finite-size DMRG algorithms
mp-dmrg-ee is renamed mp-dmrg, and is the 'default' choice
Created mp-dmrg-site as a standard 2-site algorithm
mp-dmrg-3s is the old s3e dmrg code (not yet functional)

It is probably a good idea for mp-dmrg-2site to have an RSVD option for the truncation.  Best way to do this is probably to
have a function to reshape an A-matrix over a local product basis to a matrix with the local spaces attached to the left and right.